Refactor dashboard layout and components; replace old page structure with new dashboard elements
